send messages using blocking queue
In Rust, the channel is "multi-producer, single-consumer", but in the thread pool, it actually requires "single-producer, multi-consumer", so I had to use Arc<Mutex<Receiver>> to wrap the receiver up, I believe this reduces efficiency since there will be multiple locks. So I implemented a blocking queue to optimize this.